refactor: pc와 그 외 컴포넌트 분리
