+++ frontend/src/pages/storage/components/ArticleList/ArticleList.tsx
@@ -1,0 +1,29 @@
+import styled from '@emotion/styled';
+import { Article } from '@/types/article';
+import ArticleCard from '@/pages/today/components/ArticleCard/ArticleCard';
+
+interface ArticleListProps {
+  articles: Article[];
+}
+
+export default function ArticleList({ articles }: ArticleListProps) {
+  return (
+    <Container>
+      {articles.map((article) =>
+        article ? (
+          <li key={article.articleId}>
+            <ArticleCard data={article} readVariant="badge" />
+          </li>
+        ) : null,
+      )}
+    </Container>
+  );
+}
+
+const Container = styled.ul`
+  width: 100%;
+
+  display: flex;
+  gap: 16px;
+  flex-direction: column;
+`;
